refactor: 너무 긴 컨트롤러 메서드를 쪼갬
